refactor(OMN-14331): retire models→validation import-layering back-edge in omnibase_core
Sever all 18 `omnibase_core.models.** -> omnibase_core.validation[.**]` import
back-edges (the `[importlinter:contract:core-models-no-upward]` models->validation
ignore lines) so the intra-core dependency-layering oracle (`lint-imports`) stays
green with BOTH ignore lines DELETED. Ratchet clean
(unmatched_ignore_imports_alerting=ERROR). Canary family of epic OMN-3210.

Mechanism (all model callers now import from below the models layer):
- Pure hex-color validators -> utils/util_hex_color.py (moved).
- Pure name/patch validators (is_valid_onex_name, is_valid_python_identifier,
  validate_import_path_format, validate_string_list, validate_onex_name_list,
  detect_add_remove_conflicts) -> utils/util_name_validation.py; validator_utils
  re-exports them so its public API is preserved.
- Pure topic-suffix checker -> utils/util_topic_suffix.py (model-free
  check_topic_suffix returning a NamedTuple); validation.validator_topic_suffix
  now wraps it to build ModelTopicValidationResult/ModelTopicSuffixParts (public
  API preserved). Model field-validators use check_topic_suffix directly.
- Workflow constants -> constants/constants_workflow.py (moved); MAX_DFS_ITERATIONS
  repointed to its canonical constants_field_limits home.
- runtime_ops verb loader -> utils/util_runtime_ops_verb_loader.py; a thin
  re-export kept at validation/runtime_ops_verb_loader.py for cross-repo
  consumers (omnimarket DurableEvidenceGate / node_linear_triage).
- Migration TypedDicts and the ModelProtocolInfo/ModelDuplicationInfo re-export
  bug repointed to their canonical types/ and models/validation/ sources.

dod_evidence:
- lint-imports: 4 kept, 0 broken; both models->validation ignore lines deleted.
- mypy src/omnibase_core/: 0 errors from changed files (residual redis/omnibase_spi
  errors are pre-existing, in untouched files, from missing local sibling deps).
- ruff format + check: clean.
- 2828 focused tests pass (incl. test_no_circular_imports.py), 0 failures.
- pre-commit (changed files): pass.

Refs OMN-3210
Closes OMN-14331
