+++ src/omnibase_core/models/contracts/model_capability_provided.py
@@ -16,7 +16,7 @@ Related:
 
 from pydantic import BaseModel, ConfigDict, Field, field_validator
 
-from omnibase_core.validation.validator_utils import is_valid_onex_name
+from omnibase_core.utils.util_name_validation import is_valid_onex_name
 
 __all__ = [
     "ModelCapabilityProvided",
