+++ src/omnibase_core/models/contracts/model_workflow_step.py
@@ -23,12 +23,12 @@ from omnibase_core.constants.constants_field_limits import (
     MAX_IDENTIFIER_LENGTH,
     MAX_NAME_LENGTH,
 )
-from omnibase_core.enums.enum_core_error_code import EnumCoreErrorCode
-from omnibase_core.models.errors.model_onex_error import ModelOnexError
-from omnibase_core.validation.validator_workflow_constants import (
+from omnibase_core.constants.constants_workflow import (
     MIN_TIMEOUT_MS,
     VALID_STEP_TYPES,
 )
+from omnibase_core.enums.enum_core_error_code import EnumCoreErrorCode
+from omnibase_core.models.errors.model_onex_error import ModelOnexError
 
 __all__ = ["ModelWorkflowStep", "VALID_STEP_TYPES"]
 
@@ -51,7 +51,7 @@ class ModelWorkflowStep(BaseModel):
     v1.0.4 Compliance:
         step_type MUST be one of: compute, effect, reducer, orchestrator, custom,
         parallel. The "conditional" type is reserved for v1.1+ and MUST NOT be
-        accepted in v1.0.x. See: validator_workflow_constants.VALID_STEP_TYPES
+        accepted in v1.0.x. See: constants_workflow.VALID_STEP_TYPES
     """
 
     model_config = ConfigDict(
@@ -137,7 +137,7 @@ class ModelWorkflowStep(BaseModel):
     # - Event timeout: Capped at MAX_TIMEOUT_MS (24 hours) - for longer operations
     #   See: omnibase_core/models/core/model_event_input_state.py
     # - Workflow global timeout: Configured in ModelWorkflowMetadata.timeout_ms
-    #   See: omnibase_core/validation/validator_workflow_constants.py
+    #   See: omnibase_core/constants/constants_workflow.py
     timeout_ms: int = Field(
         default=TIMEOUT_DEFAULT_MS,
         description=(
