+++ src/omnibase_core/models/contracts/subcontracts/model_request_response_instance.py
@@ -73,10 +73,9 @@ class ModelRequestResponseInstance(BaseModel):
     @classmethod
     def validate_request_topic(cls, topic: str) -> str:
         """Validate request topic suffix against ONEX naming convention."""
-        # Import here to avoid circular import at module load time
-        from omnibase_core.validation import validate_topic_suffix
+        from omnibase_core.utils.util_topic_suffix import check_topic_suffix
 
-        result = validate_topic_suffix(topic)
+        result = check_topic_suffix(topic)
         if not result.is_valid:
             raise ValueError(f"Invalid topic suffix '{topic}': {result.error}")
         return topic
