+++ src/omnibase_core/models/validation/model_migration_conflict_union.py
@@ -15,8 +15,10 @@ from pydantic import BaseModel, ConfigDict, Field, ValidationInfo, field_validat
 from omnibase_core.enums.enum_core_error_code import EnumCoreErrorCode
 from omnibase_core.enums.enum_migration_conflict_type import EnumMigrationConflictType
 from omnibase_core.models.errors.model_onex_error import ModelOnexError
-from omnibase_core.validation.validator_migration_types import (
+from omnibase_core.types.typed_dict_migration_duplicate_conflict_dict import (
     TypedDictMigrationDuplicateConflictDict,
+)
+from omnibase_core.types.typed_dict_migration_name_conflict_dict import (
     TypedDictMigrationNameConflictDict,
 )
