+++ src/omnibase_core/utils/util_workflow_executor.py
@@ -103,6 +103,14 @@ from datetime import datetime
 from typing import Literal, cast
 from uuid import UUID, uuid4
 
+from omnibase_core.constants.constants_workflow import (
+    MAX_DFS_ITERATIONS,
+    MAX_STEP_PAYLOAD_SIZE_BYTES,
+    MAX_TOTAL_PAYLOAD_SIZE_BYTES,
+    MAX_WORKFLOW_STEPS,
+    MIN_TIMEOUT_MS,
+    RESERVED_STEP_TYPES,
+)
 from omnibase_core.enums.enum_core_error_code import EnumCoreErrorCode
 from omnibase_core.enums.enum_workflow_execution import (
     EnumActionType,
@@ -129,18 +137,10 @@ from omnibase_core.models.workflow.execution.model_workflow_result_metadata impo
 )
 from omnibase_core.types.typed_dict_workflow_context import TypedDictWorkflowContext
 from omnibase_core.validation.validator_reserved_enum import validate_execution_mode
-from omnibase_core.validation.validator_workflow_constants import (
-    MAX_DFS_ITERATIONS,
-    MAX_STEP_PAYLOAD_SIZE_BYTES,
-    MAX_TOTAL_PAYLOAD_SIZE_BYTES,
-    MAX_WORKFLOW_STEPS,
-    MIN_TIMEOUT_MS,
-    RESERVED_STEP_TYPES,
-)
 
 # Note: MAX_WORKFLOW_STEPS, MAX_STEP_PAYLOAD_SIZE_BYTES, MAX_TOTAL_PAYLOAD_SIZE_BYTES
-# are imported from validator_workflow_constants.py (canonical source with memoized env var parsing).
-# See validator_workflow_constants.py module docstring for configuration details.
+# are imported from constants_workflow.py (canonical source with memoized env var parsing).
+# See constants_workflow.py module docstring for configuration details.
 
 # Module logger for workflow executor operations
 logger = logging.getLogger(__name__)
@@ -1705,7 +1705,7 @@ def _get_topological_order(
 
 
 # MAX_DFS_ITERATIONS: Resource exhaustion protection constant for DFS cycle detection.
-# Imported from validator_workflow_constants.py (canonical source).
+# Imported from constants_workflow.py (canonical source).
 # Value of 10,000 iterations supports workflows with up to ~5,000 steps
 # (worst case: each step visited twice during DFS traversal).
 # See workflow_validator.py module docstring "Security Considerations" for full documentation.
@@ -1865,8 +1865,8 @@ def verify_workflow_integrity(
 
 
 # Public API
-# NOTE: MAX_DFS_ITERATIONS is imported from validator_workflow_constants.py (canonical source).
-# Import directly from omnibase_core.validation.validator_workflow_constants for this constant.
+# NOTE: MAX_DFS_ITERATIONS is imported from constants_workflow.py (canonical source).
+# Import directly from omnibase_core.constants.constants_workflow for this constant.
 __all__ = [
     "MAX_STEP_PAYLOAD_SIZE_BYTES",
     "MAX_TOTAL_PAYLOAD_SIZE_BYTES",
